% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASICH.R
\name{lasich}
\alias{lasich}
\title{LASICH estimator}
\usage{
lasich(Sigmas, Sigma0, Omega0, ns, L = NULL, lambda1, lambda2, tol, rho,
  truegraph = NULL, initest = NULL, LD = NULL, wt = NULL, thr = TRUE)
}
\arguments{
\item{Sigmas}{(p * p * K) arrays sample covariance matrices from each group}

\item{Sigma0}{Number of case subjects}

\item{Omega0}{(p * p * K) arrays true precision matrices}

\item{ns}{vector of sample sizes for each group}

\item{L}{graph Laplacian}

\item{lambda1}{value of first tuning parameter (lambda_1)}

\item{lambda2}{value of second tuning parameter (lambda_2)}

\item{tol}{tolerance level for determining convergence of LASICH}

\item{rho}{tuning parameter for ADMM algorithm}

\item{truegraph}{a list containing true graphs (to return performance evaluation metrics)}

\item{initest}{initial estimator for warm start}

\item{LD}{a diagonal matrix added to L to improve computational efficiency}

\item{wt}{weight for likelihood; NULL = nk/n, 1 = (1,...,1)???}

\item{thr}{whether to threshold the sample cov matrix to find connected components of the graph}

\item{Sigma0}{(p * p * K) arrays true covariance matrices}
}
\description{
Calculates sensitivity and specificity of estimating K precision matrices using LASICH. 
Returns precision matrix estimates, Frobenius norm errors, and operating characteristics.
}
